Add files via upload
SHAP is a handy tool to determine how certain variables may impact our option's price. Looking at the SHAP diagram, it is important to note that the X-axis, impact on value, will remain the same as in our case - while the variables will be different, as this is merely an example of how we can use this library to our advantage in determining what variables impact options prices greater than others.